Added function to add release and mark as wanted by MusicBrainz release id.
This required a few changes in mb.py - the getRelease request is now a
big heavier as it is pulling in the ReleaseGroup and Artist information
as well through include objects. I think this should always be done to
save repeated requests anyway. A "light" version of the artist is added
with status="Manual" and no extra releases/releasegroups are found.