New Feature - Exclude Skipped Albums In Total in Have Counts
Added a new option to the configuration page called "Exclude Skipped
Albums In Total in Have Counts". When ticked, the total "Have" count for
an artist will only include albums marked as Downloaded or Wanted.

The totaltracks for the artist will not include skipped albums.

A "Force Update Active Artists" is required when the artist already
exists in the headphones database.