Resolve library import edge case
When adding a library, it is possible the single space character (' ') is added as an artist. The Musicbrainz function _do_mb_search is not capable of handling this, and throws a ValueError, as no query is provided. This checks if an artist fits this situation, and continues with the next possible artist.