Fix for Duplicate Torrent and "Stuck in search for wanted"
Checks for KeyError looking for key 'torrent-added' to make sure that
the success result is not caused by something other than a torrent
being added, for example a duplicate torrent with {u'arguments':
{u'torrent-duplicate': … instead of {u'arguments': {u'torrent-added':…

Removed loop in getTorrentFolder to allow calling searher.py to
continue searching for next torrent. Earlier the thread was stuck
waiting for the percentage to become more than 0 percent complete,
which may frequently never happen if no seeders are online. I could not
figure out why the loop was needed since the torrent_folder_name is set
the line above the loop. Have not had any problems yet forcing me to
retry every five seconds like in the loop.