Fix when two instances of torrent_folder_name are different
The two instances of torrent_folder_name can be different - if the
torrent hasn't started then torrent_folder_name will be the torrent
name which will then be overridden with the torrent folder when the
torrent has started.

Most of the time these will be the same but sometimes they're not and
the post processor does not work without this fix.