Change pushover notification title
Makes more sense to set pushover notification titles as "Headphones" rather than "Download and Postprocessing completed" - which may be too long (gets truncated on android app depending on screen resolution) and not descriptive when you have lots of other notifications coming in. If that verbiage is desired, it would be better in the "pushmessage" body. 

This may apply to the other notification types as well, but not sure as I don't use them.