Postgres Support: Core API
Implement some core api bits around postgresql support.  For this to
work, binding formats will have to be changed throughout the rest of the
source, hoping this will work for mitigating that.  Future commits will
do the binding parameter changes, and alter some of the fetches to be
consistent with psql's and sqlite's expectations.