fixup migrations for psql
psql rollback semantics are different than sqlite's, so we need to
rollback after test 'failures' in the DB migration phase.

use a flag (a table/timestamp in this case) to detect when we've done
the artists lowercase transform as well, so we don't get bit if the
first artistname collation happens to be lowercased. (unlikely on most
libraries, but definitely possible).